Creating a marathon-template executable
1. Moving things from scripts to app, makes it look a little cleaner
2. Extracted the marathon main out to marathon-template executable
